feat: fix all screen and layout
