+++ app/src/main/java/com/bangkit/turtlify/ui/about/AboutActivity.kt
@@ -2,6 +2,8 @@ package com.bangkit.turtlify.ui.about
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
+import com.bangkit.turtlify.R
 import com.bangkit.turtlify.databinding.ActivityAboutBinding
 import com.bangkit.turtlify.databinding.ActivitySettingsBinding
 
@@ -12,5 +14,20 @@ class AboutActivity : AppCompatActivity(){
         binding = ActivityAboutBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        setupView()
+    }
+
+    private fun setupView() {
+        val toolbar = findViewById<Toolbar>(R.id.my_toolbar)
+        setSupportActionBar(toolbar)
+        supportActionBar?.apply {
+            title = getString(R.string.title_activity_about)
+            setDisplayHomeAsUpEnabled(true)
+        }
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        finish()
+        return true
     }
 }
