+++ app/src/main/java/com/bangkit/turtlify/ui/faq/FaqActivity.kt
@@ -2,7 +2,9 @@ package com.bangkit.turtlify.ui.faq
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.bangkit.turtlify.R
 import com.bangkit.turtlify.data.model.FaqData
 import com.bangkit.turtlify.databinding.ActivityFaqBinding
 import com.bangkit.turtlify.databinding.ActivitySettingsBinding
@@ -14,7 +16,16 @@ class FaqActivity : AppCompatActivity() {
         binding = ActivityFaqBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        setupView()
         setupRecyclerView()
+    }
+    private fun setupView() {
+        val toolbar = findViewById<Toolbar>(R.id.my_toolbar)
+        setSupportActionBar(toolbar)
+        supportActionBar?.apply {
+            title = getString(R.string.title_activity_faq)
+            setDisplayHomeAsUpEnabled(true)
+        }
     }
 
     private fun setupRecyclerView() {
@@ -23,5 +34,10 @@ class FaqActivity : AppCompatActivity() {
 
         binding.rvItemFaq.layoutManager = LinearLayoutManager(this)
         binding.rvItemFaq.adapter = faqAdapter
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        finish()
+        return true
     }
 }
