+++ app/src/main/java/com/bangkit/turtlify/ui/suggestion/SuggestionActivity.kt
@@ -2,6 +2,8 @@ package com.bangkit.turtlify.ui.suggestion
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
+import com.bangkit.turtlify.R
 import com.bangkit.turtlify.databinding.ActivitySuggestionBinding
 
 class SuggestionActivity : AppCompatActivity() {
@@ -10,5 +12,21 @@ class SuggestionActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActivitySuggestionBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        setupView()
+    }
+
+    private fun setupView() {
+        val toolbar = findViewById<Toolbar>(R.id.my_toolbar)
+        setSupportActionBar(toolbar)
+        supportActionBar?.apply {
+            title = getString(R.string.title_activity_suggestion)
+            setDisplayHomeAsUpEnabled(true)
+        }
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        finish()
+        return true
     }
 }
