+++ app/src/main/res/layout/activity_camera.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activities.CameraActivity">
+    tools:context=".ui.camera.CameraActivity">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/viewFinder"
