+++ app/src/main/res/layout/activity_settings.xml
@@ -1,38 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
+<androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="16dp">
+    android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/tvSettingsTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Settings"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="16dp"/>
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/my_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:layout_alignParentTop="true"
+        android:elevation="4dp"
+        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
     <LinearLayout
         android:id="@+id/llSettingsMenu"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@id/tvSettingsTitle"
+        android:layout_below="@id/my_toolbar"
         android:orientation="vertical"
-        android:layout_marginTop="16dp">
+        app:layout_constraintTop_toBottomOf="@+id/my_toolbar"
+        tools:layout_editor_absoluteX="16dp"
+        android:padding="16dp">
 
         <!-- Language Menu Item -->
         <TextView
             android:id="@+id/tvLanguage"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="@string/language"
-            android:textSize="18sp"
             android:clickable="true"
             android:padding="8dp"
+            android:text="@string/language"
+            android:textSize="18sp"
             app:drawableEndCompat="@drawable/ic_arrow_right" />
 
         <!-- FAQ Menu Item -->
@@ -40,11 +44,11 @@
             android:id="@+id/tvFaq"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="@string/faq"
-            android:textSize="18sp"
+            android:layout_marginTop="8dp"
             android:clickable="true"
             android:padding="8dp"
-            android:layout_marginTop="8dp"
+            android:text="@string/faq"
+            android:textSize="18sp"
             app:drawableEndCompat="@drawable/ic_arrow_right" />
 
         <!-- Suggestion Menu Item -->
@@ -52,11 +56,11 @@
             android:id="@+id/tvSuggestion"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="@string/suggestion"
-            android:textSize="18sp"
+            android:layout_marginTop="8dp"
             android:clickable="true"
             android:padding="8dp"
-            android:layout_marginTop="8dp"
+            android:text="@string/suggestion"
+            android:textSize="18sp"
             app:drawableEndCompat="@drawable/ic_arrow_right" />
 
         <!-- About Menu Item -->
@@ -64,22 +68,22 @@
             android:id="@+id/tvAbout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="@string/about"
-            android:textSize="18sp"
+            android:layout_marginTop="8dp"
             android:clickable="true"
             android:padding="8dp"
-            android:layout_marginTop="8dp"
+            android:text="@string/about"
+            android:textSize="18sp"
             app:drawableEndCompat="@drawable/ic_arrow_right" />
 
         <TextView
             android:id="@+id/tvShare"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="@string/share"
-            android:textSize="18sp"
+            android:layout_marginTop="8dp"
             android:clickable="true"
             android:padding="8dp"
-            android:layout_marginTop="8dp"
+            android:text="@string/share"
+            android:textSize="18sp"
             app:drawableEndCompat="@drawable/ic_arrow_right" />
 
         <!-- Dark Mode Switch -->
@@ -88,23 +92,23 @@
             android:id="@+id/llDarkMode"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:layout_marginTop="8dp">
+            android:layout_marginTop="8dp"
+            android:orientation="horizontal">
 
             <TextView
                 android:id="@+id/tvDarkMode"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
+                android:padding="8dp"
                 android:text="@string/dark_mode"
-                android:textSize="18sp"
-                android:padding="8dp"/>
+                android:textSize="18sp" />
 
             <Switch
                 android:id="@+id/switchDarkMode"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content"/>
+                android:layout_height="wrap_content" />
         </LinearLayout>
 
     </LinearLayout>
-</RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
