+++ app/src/main/res/values/preloaded_fonts.xml
@@ -17,6 +17,6 @@
         <item>@font/poppins_medium_italic</item>
         <item>@font/poppins_semibold</item>
         <item>@font/poppins_semibold_italic</item>
-        <item>@font/poppins_thin</item>
+<!--        <item>@font/poppins_thin</item>-->
     </array>
 </resources>
